#2437 split phoebus opi integration into stand alone plugin
