+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.opibuilder/META-INF/MANIFEST.MF
@@ -1,0 +1,14 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Phoebus OPIbuilder Integration
+Bundle-SymbolicName: org.csstudio.phoebus.integration.opibuilder;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Activator: org.csstudio.phoebus.integration.opibuilder.Activator
+Require-Bundle: org.eclipse.core.runtime,
+ org.eclipse.ui,
+ org.csstudio.ui.menu,
+ org.csstudio.ui.util,
+ org.csstudio.csdata,
+ org.csstudio.phoebus.integration
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Bundle-ActivationPolicy: lazy
