+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.opibuilder/build.properties
@@ -1,0 +1,8 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.xml,\
+               preferences.ini,\
+               pom.xml,\
+               icons/
