+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.opibuilder/icons/OPIBuilder.png
