+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.opibuilder/icons/OPIRunner.png
