+++ applications/phoebus/phoebus-plugins/org.csstudio.phoebus.integration.opibuilder/pom.xml
@@ -1,0 +1,24 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <artifactId>org.csstudio.phoebus.integration.opibuilder</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
+  <build>
+    <sourceDirectory>src</sourceDirectory>
+    <plugins>
+      <plugin>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>3.7.0</version>
+        <configuration>
+          <source>1.8</source>
+          <target>1.8</target>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+  <parent>
+    <groupId>org.csstudio</groupId>
+    <artifactId>phoebus-plugins</artifactId>
+    <version>4.2.0-SNAPSHOT</version>
+  </parent>
+  <packaging>eclipse-plugin</packaging>
+</project>
